+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - [#275](https://github.com/Datatamer/unify-client-python/issues/275) Create a category with a category spec
   - [#273](https://github.com/Datatamer/unify-client-python/issues/273) Attribute type spec to allow for attribute creation
   - [#219](https://github.com/Datatamer/tamr-client/issues/219) Delete a resource from collection.
+  - [#277](https://github.com/Datatamer/unify-client-python/issues/277) Attribute mapping spec
 
   **BUG FIXES**
   - [#235](https://github.com/Datatamer/unify-client-python/issues/235) Making `AttributeCollection` retrieve attributes directly instead of by streaming